[Fix]: `RunQ.listRunQ` returns `null` data
`RunQ.listRunQ` returned a `null` object reference in some cases.
We changed it to return an empty , 0 sized String array insted.
This doesn’t break existing logic, but prevents a possible
`NullPointerException`.

Example: submitting a run with a non existent benchmark name caused a
`NullPointerException` on the `CLIServlet` side. Now it behaves as
expected, returning HTTP status code NOT FOUND.